提案されたトレーニングメニューという表記を削除し、ファイルの名前にMenを追加
